requirements.txt
Flask==1.1.2
netifaces==0.10.9
requests==2.23.0
uptime==3.0.1
grpcio==1.32.0